Refactor class/file location
