+++ ActiveDirectoryQuerier.Tests/ADCommandParametersTests.cs
@@ -1,4 +1,5 @@
 ﻿using System.Management.Automation.Runspaces;
+using ActiveDirectoryQuerier.ActiveDirectory;
 using ActiveDirectoryQuerier.PowerShell;
 
 namespace ActiveDirectoryQuerier.Tests;
