+++ ActiveDirectoryQuerier/ActiveDirectoryQuerier.csproj
@@ -26,7 +26,6 @@
 	</ItemGroup>
 
 	<ItemGroup>
-	  <Folder Include="ActiveDirectory\" />
 	  <Folder Include="ViewModels\" />
 	</ItemGroup>
